% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{instruments}
\alias{instruments}
\title{Get Instruments Data}
\usage{
instruments(assets = "stocks")
}
\arguments{
\item{assets}{Type of assets to query, by default 'stocks'.}
}
\value{
A data frame containing instrument data, or NULL if the request fails.
}
\description{
Retrieves instruments data filtered by asset type. This function
uses an API token stored in an environment variable named 'LIMEX_API_TOKEN'.
}
\examples{
\dontrun{
  instruments_data <- instruments(assets = 'stocks')
}
}
