% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limex_token.R
\name{limex_token}
\alias{limex_token}
\title{Set API Token for Limex}
\usage{
limex_token(token)
}
\arguments{
\item{token}{API token for the Limex service.}
}
\value{
Invisible NULL, side-effect function setting an environment variable.
}
\description{
Sets the API token for the Limex service by storing it in an
environment variable for the current R session. This token will be used by
other functions in the package to authenticate API requests.
}
\examples{
\dontrun{
  limex_token("your_personal_token_here")
}
}
