% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Retrieve Models Data from Limex API}
\usage{
models(vendor = "boosted")
}
\arguments{
\item{vendor}{The name of the vendor for which models data is requested.}
}
\value{
A data frame containing the models data if the request is successful; NULL otherwise.
}
\description{
Fetches data about models from the Limex API for a specified vendor.
It uses an API token that is stored in the 'LIMEX_API_TOKEN' environment variable.
}
\examples{
\dontrun{
  models_data <- models(vendor = "boosted")
}
}
